findContainingComponentForPathInProjectContents moved to element-template-utils
